чёто

+++ src/main/java/com/nikkin/devicesdb/Dto/Identifiable.java
@@ -1,5 +1,5 @@
 package com.nikkin.devicesdb.Dto;
 
 public interface Identifiable {
-    Long id();
+    Integer id();
 }
